added extended log to indeitfy issue with
random missing of certain variables in the zarr output
for gfs such as clwmr, the process_dataframe was edit
to avoid missing acpcp. However the clwmr missing issue
is persisting and differing based on dates, such as
on 20240825, 20230825, but it is availble in dates 20210825
and 20220825.

With logs, the issue stem from build_idx_grib_mapping
and there is variable duplicates identified in validate
part of the function, investigating.